// api/types.gen.ts
export type TaskRunInputInput = {
  /**
   * Task specification.
   */
  task_spec?: TaskSpec | null
  /**
   * Input
   * Input to the task, either text or a JSON object.
   */
  input:
    | string
    | {
        [key: string]: unknown
      }
  /**
   * Processor
   * Processor to use for the task.
   */
  processor: string
  /**
   * Metadata
   * User-provided metadata stored with the run. Keys and values must be strings with a maximum length of 16 and 512 characters respectively.
   */
  metadata?: {
    [key: string]: string | number | number | boolean
  } | null
}
// ...
export type TaskSpec = {
  /**
   * Output Schema
   * JSON schema or text fully describing the desired output from the task. Descriptions of output fields will determine the form and content of the response. A bare string is equivalent to a text schema with the same description.
   */
  output_schema: JsonSchema | TextSchema | string
  /**
   * Input Schema
   * Optional JSON schema or text description of expected input to the task. A bare string is equivalent to a text schema with the same description.
   */
  input_schema?: JsonSchema | TextSchema | string | null
}
// ...
export type TasksRunsPostV1TasksRunsPostData = {
  body: TaskRunInputInput
  path?: never
  query?: never
  url: '/v1/tasks/runs'
}
// ...
export type TasksRunsGetV1TasksRunsRunIdGetData = {
  body?: never
  path: {
    /**
     * Run Id
     */
    run_id: string
  }
  query?: never
  url: '/v1/tasks/runs/{run_id}'
}
// ...
export type TasksRunsInputGetV1TasksRunsRunIdInputGetData = {
  body?: never
  path: {
    /**
     * Run Id
     */
    run_id: string
  }
  query?: never
  url: '/v1/tasks/runs/{run_id}/input'
}
// ...
export type TasksRunsResultGetV1TasksRunsRunIdResultGetData = {
  body?: never
  path: {
    /**
     * Run Id
     */
    run_id: string
  }
  query?: {
    /**
     * Timeout
     */
    timeout?: number
  }
  url: '/v1/tasks/runs/{run_id}/result'
}